+++ pymeos/pymeos/aggregators/point_aggregators.py
@@ -6,6 +6,13 @@ from ..main import TPoint
 
 
 class TemporalPointExtentAggregator(BaseAggregator[TPoint, STBox]):
+    """
+    Spatiotemporal extent of aggregated temporal points, i.e. smallest :class:`~pymeos.time.boxes.STBox` that
+    includes all aggregated temporal numbers.
+
+    MEOS Functions:
+        tpoint_extent_transfn, temporal_tagg_finalfn
+    """
     _add_function = tpoint_extent_transfn
 
     @classmethod